Layout bricks for game and properly free objects
I worked on the bricks to lay them out properly in a grid.
I did it line by line instead of via a loop since there are only a fixed amount of
bricks. I also edited the ArkaObject so that I could implement proper freeing resources.
After 'destroy()' is called on an object it is marked to be freed and its slot in the game's
object buffer is set to null. Then the JS runtime would free the object from memory.