Fix refresh rate inconsistency and more
The ball should now move at the same speed regardless
of screen refresh rate. The rendering canvas should now auto resize
upon resizing the window. There was some strange behavior with the ball going
out of bounds. While it corrected itself, it took while to return to the main screen.
I implemented a condition that would reset the ball's position if it ever went too far.
Was also going to implement pausing and I had it done. I decided to remove it for now since it could
make the game too easy.